build(cmake): lower CMAKE_C_STANDARD to 11
per
https://cmake.org/cmake/help/latest/prop_gbl/CMAKE_C_KNOWN_FEATURES.html,
c_std_17 was introduced in CMake 3.21, while the required CMake
version is 3.20. which is packaged by RHEL/RockyLinux 9, so it would be
nice if we could build with co_context on these distro releases using
the shipped CMake.

so, in this change, CMAKE_C_STANDARD is changed from 17 to 11. this
change should have minimum impact, as co_context is a C++20 project
which does not contain C source code.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>